requirements.txt
pandas
numpy
pyyaml
python-dotenv
requests
solana>=0.36
solders>=0.26.0
base58
